Added code to collect all of an application's defaults into a Registry.  This could be used to dump an input file template, for example.
The code finds all possible components using brute
force:  for each facility, it finds every .odb file with
a matching factory method, and calls it (!).  It would
be more graceful to use metadata, but design choices in
Pyre as well as limitations in Python interplay to make
such an approach difficult, if not impossible.