Fix-up sys.path so that e.g., "import pyre" works.
In the application, a single "import pythia" is needed
somewhere (e.g., Foo/__init__.py, or the top-level script).
If this were C++, "import pythia" would be equivalent to

    using namespace pythia;

It would be nice if pkg_resources provided an API for
this:

    from pkg_resources import use
    use('pythia', __name__)

Or maybe a language extension? :-)

The difference being that this would modify __path__
instead of sys.path for modules.  At least, I think that
would work.  I tried to tweak e.g., pyre.__path__ from
within pythia, but it lead to an infinite regress no
matter how I did it.  It might require a custom loader
(Pmw has one), I'm not sure.