Removed hard-coded "-np" argument-list code for 'mpirun'. Instead, apply macro expansion to launcher.command.  The default is
    mpirun -np ${nodes}

but on TACC's new Lonestar system, for example, one can
configure this as follows:

    [CitcomS.launcher]
    command = ibrun

This will cause the "-np" arguments to be omitted entirely,
as required.  Also removed the 'extra' property, which is
certainly redundant now (if it wasn't already before).

Also fixed another undefined variable bug, this time in
SchedulerLSF.  Once again, it was an an error path that
isn't always taken.  You gotta love Python...