Added support for PBS.  This included adding 'os.environ' to the MPI launcher macro substitutions, so that one can reference PBS_NODEFILE as follows:
    [CitcomS.launcher]
    command = mpirun -np ${nodes} -machinefile ${PBS_NODEFILE}

Nuked all references to FileLocking (actually there was only one, in
CodecODB).  This is the annoying code that prevented Pyre from working
on NFS-mounted filesystems (!).  The locking seems to serve no
purpose, anyway.

Renamed 'scheduler.batch-command' to simply 'scheduler.command'.  The
'batch-' was an anachonism: it was to avoid a name conflict with the
'command' property used for 'mpirun', back when Launcher and Scheduler
were the same thing.