Made 'journal' configurable in the traditional way.  For example:
    myscript --journal.debug.myscript --journal.device=file

will turn-on 'myscript' debug log messages and send all journal
messages to a file.  Note: this is all done via smoke and mirrors.
Under CIG-Pyre, 'journal' is not an ordinary facility, because it must
be initialized early.  See the log comment for r5356.

Made Application.run() fire-up the new start-up procedure, so that the
official Pyre test cases and examples will behave identically to newer
stuff that goes through pyre.applications.start().

Finally, changed the default for "excepthook.call-pdb" to 'False'.
The call_pdb feature seems to be broken in the latest version of
IPython (0.7.3).  Besides, landing in pdb after an exception is
arguably unexpected behavior.