Issue5: Implemented simple Bash "programmable completion" for Pyre apps.  Add the following to your .bashrc to use:
    complete -C "pyreapp -c" pyreapp

where 'pyreapp' is any Pyre application, e.g.,

    complete -C "citcoms -c" citcoms

~~~ Benefits ~~~

Say you don't remember how to set the time limit for a job.  Type

    citcoms <TAB>

A list of all top-level properties and facilities is presented (plus
all .cfg and .pml files in the current working directory).

Hmmm.  Perhaps it is under 'job'.  So you continue typing:

    citcoms --job.<TAB>

A list of the properties and facilities for 'job' appears.  Ah, there
it is -- "job.walltime".  But you are lazy, so you just type a 'w' and
hit <TAB> again:

    citcoms --job.w<TAB>

And the shell completes "walltime" for you :-)

~~~ Limitations ~~~

Pyre will not help you with trait values:

    myapp --foo=<TAB>

will do nothing.  But in the future, it could print a list of valid
values for 'foo' in certain cases.

Currently, tab-completion does not work for apps which override the
default CommandlineParser.  (There is exactly one such app: PyLith
v0.8.)

~~~ Bugs ~~~

If you override a facility setting on the command line, the
completions for traits of the corresponding component will be
wrong.  E.g.,

    myapp --solver=xyz --solver.<TAB>

will show you the traits for the solver set by default or in a
configuration file -- not necessarily those of the 'xyz' solver.