Merged Django v0.95.1 into Opal.
~~~~

"Any barrier that exists between the user and some part of the system
will eventually be a barrier to creative expression. Any part of the
system that cannot be changed or that is not sufficiently general is a
likely source of impediment. If one part of the system works
differently from all the rest, that part will require additional
effort to control. Such an added burden may detract from the final
result and will inhibit future endeavors in that area."

--Dan Ingalls
 "Design Principles Behind Smalltalk"
 
http://users.ipa.net/~dwighth/smalltalk/byte_aug81/design_principles_behind_smalltalk.html