Make some long-overdue improvements to the core framework, to address the challenges presented by PyLith-0.8 and SPECFEM.
~~~ _validate() ~~~

First and foremost: Components can now define a _validate() method
which performs complex validation, and generates error messages which
are every bit as pretty as those produced by the framework itself
(rather than just bombing with an exception).  "Complex" validation is
defined as any validation checks involving two or more properties, and
which therefore cannot by handled by a simple validator.  For example:

    def _validate(self, context):
        Component._validate(self, context)

        # convert to degrees
        from pyre.units.angle import deg
        ANGULAR_WIDTH_XI_IN_DEGREES  = self.inventory.angular_width_xi / deg

        # this MUST be 90 degrees for two chunks or more to match geometrically
        if (self.inventory.NCHUNKS > 1 and
            ANGULAR_WIDTH_XI_IN_DEGREES != 90.0):
            context.error(ValueError("'angular-width-xi' must be 90 degrees for more than one chunk"),
                          items=[self.metainventory.angular_width_xi])

Note that the error is not raised, but reported to the 'context'
argument.  This allows the application to report multiple user errors,
as a compiler would (the gold standard to which CIG-Pyre aspires) --
rather than bombing on the first one (as CACR-Pyre often does).  The
'items' keyword argument allows the framework to report the location
from which the bogus 'angular-width-xi' value orginated (as the
framework would for its own error messages).

The _validate() method fires immediately before _configure().


~~~ MetaInventory ~~~

Notice that the above code uses something called 'metainventory'.
This was invented simply to make writing _validate() easier.  However,
it has other useful applications as well:

    # dump all inventory items, and where they came from
    for item in self.metainventory:
        print item.attr, item.name, item.value, item.locator

(This was possible all along, it just wasn't so easy.)  The addition
of 'attr' to class Trait (and the metainventory items) makes it easy
to automatically transfer all inventory items to ordinary instance
variables:

    def _configure(self):
        Component._configure(self)
        for item in self.metainventory.iteritems(MyComponent.Inventory):
            setattr(self, item.attr, item.value)

(One can achieve the same effect by declaring traits as class
attributes, instead of nested inside Inventory; see r4743.  But some
people like having a nested Inventory class -- go figure.)  The class
argument to 'iteritems' is significant: if not given, all inventory
items (including those declared in base and derived components) are
enumerated.


~~~ When 'default' is 'None', revisited ~~~

Facilities are now consistent with properties, allowing components to
specify optional plug-ins:

    plugin = pyre.facility("plug-in", default=None)

As with properties, a value of None means the facility was not set:

    if self.inventory.plugin is not None:
        # a plugin was plugged in!
    else:
        # no 'plug-in' was specified

As before, it is an error to omit both the default and the factory
method:

    plugin = pyre.facility("plug-in") # error

Revisiting r6258, inputFile and outputFile now have the original CACR
semantics when 'default' is not specified; e.g.:

    import pyre.inventory as pyre

    # defaults to 'stdout'
    dump = pyre.outputFile("dump")

    # no default; must check for 'None'
    winkler = pyre.outputFile("winkler", default=None)


~~~ Unknown components, unrecognized properties ~~~

Unknown components and unrecognized properties are now accumlated into
a registry.  In the future, this might make the ComponentHarness
implementation robust and beautiful.  The immediate benefit is that
unknown component errors are more informative:

 >> myerrors.cfg:3:
 -- pyre.inventory(error)
 -- asdljfasf.asdfljka.x <- '2'
 -- unknown component 'asdljfasf.asdfljka'

Instead of listing all the unknown components on a single line, Pyre
now reports each occurrance individually, and where it came from.


~~~ misc. ~~~

Exceptions thrown during Component instantiation are now caught.
SystemExit exceptions were mistakenly being caught by class Inventory
under Python 2.4 and earlier, now they aren't.

I cleaned up Locators a little bit, esp. how they interface with
Journal.  Added StackTraceLocator.