Added the ${td} and ${basename} macros for .cfg files.
~~~ ${td} ~~~

The ${td} macro expands to the directory pathname of the .cfg file
("td" == "this directory").  Within a .cfg file, ${td} allows one to
reference other files relative to the .cfg file itself:


    # PyLith
    [pylith3d]
    bcInputFile = ${td}/splittest.bc


    # CitcomS
    [CitcomS.solver.mesher]
    # Read uneven mesh point coordinates from 'coor.dat'.
    coor = on
    coor_file = ${td}/coor.dat


This allows examples to work from anywhere, regardless of the current
working directory:

    citcoms Cookbook4/cookbook4.cfg

The output files will appear in the working directory as always, but
"${td}/coor.dat" expands to "Cookbook4/coor.dat", so that the input
file "coor.dat" is found alongside "cookbook4.cfg".

This works for all .cfg files; not merely those named on the command
line.  So, within ~/.pyre/pylith3d/pylith3d.cfg, ${td} expands to
"$HOME/.pyre/pylith3d".

~~~ ${basename} ~~~

The ${basename} macro expands to the basename of the .cfg file -- the
pathname stripped of the directory component and the '.cfg' extension
(e.g., "cookbook4").  With input files, it would commonly be used in
conjuction with ${td}, but with output files, it should probably be
used alone:

    # splittest.cfg
    [pylith3d]
    asciiOutputFile = ${basename}.ascii
    bcInputFile = ${td}/${basename}.bc

The advantage of this is that it allows an entire set of input/output
files to be toggled simply by copying or renaming the .cfg file:

    cp splittest.cfg mysplittest.cfg
    pylith3dapp.py mysplittest.cfg

Now PyLith would read "mysplittest.bc" and write "mysplittest.ascii",
simply by virtue of the new .cfg filename.

The ${basename} macro is perhaps a little goofy.  The goal is to
capture most of spirit of PyLith's ad-hoc ${fileRoot} macro, but in a
way that can implemented cleanly and simply within the Pyre framework.