Expand macros in defaults:
    foo = pyre.inventory.str("foo", default="${bar}")

This change has the side-effect of making file property defaults more
consistent:

    input = pyre.inventory.inputFile("input", default="default.txt")

This will now fail with a pretty error message during the
configuration phase if "default.txt" does not exist.  Before, it would
die with an exception when (and if) the 'input' inventory item was
accessed; and only print a pretty error message for a missing file if
the user explicitly set the "input" property (overriding the default).

Defaults take a different path through the framework than
user-specified values do.  In most cases, a bogus default value is a
bug in the application, so handling defaults in a consistent fashion
isn't a big deal.  InputFile and OutputFile, however, are different --
consequently, their behavior has always been slightly broken.
Fortunately, no one (other than me) has been using them (although this
will change).