Added some handy built-in macros:
    ${job.name} : LSF/PBS job name specified using "job.name"
    ${nodes}    : number of machine nodes
    ${rank}     : result of MPI_Comm_rank(MPI_COMM_WORLD) 

These are only expanded when the job is running (i.e., they are not
expanded on the head/login node).  The ${rank} macro is only expanded
when running in parallel under 'mpirun'.

These built-in macros are the low-hanging fruit.  There are more
built-in macros to come...

(Note that Launcher's specialized macro processor for
'launcher.command' has now been replaced by the general-purpose macro
facility.)