Complain about typos in the app name in .cfg files -- e.g.:
[citcoms]
steps = 5

Previously, this option would be silently ignored, since the user
wrote "citcoms" instead of "CitcomS".

These are diagnosed as "unknown components", which means they are not
reported when --typos=relaxed, they are reported as warnings when
--typos=strict (the default), and they are reported as errors when
--typos=pedantic.

This change assumes that .cfg/.pml files aren't shared across
applications, which seems to be the case.